// services/openwork/apps/app/src/app/types.ts
export type StartupPreference = "local" | "server";
// ...
export type WorkspaceDisplay = WorkspaceInfo & {
  name: string;
};